+++ database/migrations/2020_10_07_075949_create_offers_table.php
@@ -18,7 +18,7 @@ class CreateOffersTable extends Migration
             $table->string('code');
             $table->date('date');
             $table->string('validate')->default(\App\Offers::VALIDATE);
-            $table->integer('offer')->unsigned();
+            $table->unsignedBigInteger('offer')->unsigned();
             $table->timestamps();
         });
     }
