+++ database/migrations/2020_10_07_080049_create_questions_table.php
@@ -16,7 +16,7 @@ class CreateQuestionsTable extends Migration
         Schema::create('questions', function (Blueprint $table) {
             $table->id();
             $table->string('question');
-            $table->integer('question_type_id')->unsigned();
+            $table->unsignedBigInteger('question_type_id')->unsigned();
             $table->timestamps();
         });
         Schema::table('questions' ,  function (BluePrint $table){
